Add OpenGate, ClosedGate
After each gravity tick, if an OpenGate was occupied and now isn't, it
becomes a solid ClosedGate.

(The list of occupied OpenGates is added to before and after initial
movement, and after portals activate.)

This is different to the behaviour of portal activation - for example,
you can push different objects past an OpenGate without it closing, and
can even pull a snake tail away from an OpenGate, but have an object
fall on it in the first gravity tick, without it closing. (If it takes
more than one gravity tick, it will close.)

I tested all portal/block/snake interactions me and my brother Nick
could think of, and they all work the way Nick wants them to.

(Also fixed a bug with foam not being deleted if you portal onto it, and
changed some minor syntax things to be more like the codebase)